FailOverCacheAsyncBridge: fix adding entries
addEntries checks the throttling based on locs.size() which is (at the
moment) of fixed size whereas it should actually use num_locs. As a
consequence we're too often in throttling mode.
Next up in the call chain - if we're not throttling - is
addEntry (needlessly checks for throttling again, so that is
removed). There the code also triggers an async writeout if the newOnes
vector has reached a certain size. This should however be performed in
any case, i.e. even if addEntry is not even called as we're throttling.

Pt. 1 of the fixes for OVS-3410.