Paper over ALBA uploads eventually finishing despite reporting timeouts
With ALBA it's possible (and has been observed as well), that object
uploads are reported as failed (with "invalid inputstream", indicating a
timeout on the control channel), whereas the actual upload still
continues in the background.
VolumeDriver's backend code (BackendInterface) will retry after a
timeout, but if the caller requests 'OverwriteObject::F' then the retry
will fail in the above scenario as the object now exists.

This makeshift fix changes the scrubbing code to permit uploads with
overwrite after checking that the object does not exist. This opens a
window for race conditions which should however be prevented by the
namespace locking (at least races of different scrubber instances).

Addresses OVS-3615.