youtils::Logger: support for multiple sinks
This also changes the command line options supported by volumedriver
binaries: --logfile is gone in favour of --logsink, which
can be specified multiple times to allow logging to multiple files
and/or the console ("--logsink=console:").
Omitting --logsink retains the current behaviour of using the console
for logging.

Part of OVS-4088.