Volume: revert the rwlock back to fungi::RWLock
Microbenchmarks show that the fungi::RWLock prefers readers and offers
performance benefits in the absence of writers. Since the rwlock is now
only taken in exclusive mode by management actions, this behaviour is
desirable (vs. boost's shared mutex which is fair towards writers but
does not perform as well in the absence of writers).

Part of OVS-4283.