DTL client: disallow dynamic reconfiguration of qdepth
The code does not handle reconfigurations correctly as there are race
conditions and assertions that will trigger on reconfiguration. For the
time being dynamically picking up changes to the global dtl_queue_depth
option will not be picked up dynamically but only when reconfiguring the
Volume's DTL client.

Fixes OVS-4318.