AmqpEventPublisher: revise internal locking
There used to be a single lock to serialize reconfiguration as well as
access to the underlying SimplAmqpClient Channel. Calling a method of
the latter was however observed to block indefinitely (waiting for a
message from the RabbitMQ server) which is rather undesirable behaviour
while holding a lock.
Since there's no way to reliably time out all calls of Channel methods,
the locking was revised to at least avoid reconfigurations getting stuck
on this.
That still means that event submissions can get stuck but this was
addressed earlier on already by moving the actual submission to a thread
that consumes events from a bounded queue - cf. b4e8e6b1d3

Addresses OVS-4527