filesystem: serialize unaligned writes
As voldrv performs read/modify/write for unaligned (to the volume's
cluster size) writes, these have to be serialized. So far this was
implicitly provided by the interfaces on top of it but the newly
introduced edge exposes this as it allows concurrent writes.

Hotfix for https://github.com/openvstorage/volumedriver/issues/48