ClusterRegistry: fix backward compatibility of arakoon asserts
When modifying the state of nodes in the registry the code asserts on
the previous state of the registry to detect / deal with concurrent
updates. The previous state of the registry was however deserialized and
serialized again for the assertion, which in case of newly added fields
to the serialized format caused the assertion to fail.

Addresses https://github.com/openvstorage/volumedriver/issues/129