Python API: add `req_timeout_secs` arg to API calls
This allows to specify a client timeout per method call in
`{Local,}StorageRouterClient` instances. A call that times out on the
client side is *not* aborted on the server side and will still run to
completion there.

Default value: `None` -> maintains current behaviour of using the
timeout configured on the instance itself.

Addresses https://github.com/openvstorage/volumedriver/issues/128