youtils/test/CapnProtoTest
A bunch of tests of Capn'Proto's fitness for the async DTL purposes.

The RPC tests are compiled out as they have the unpleasant side effect
of crashing gdb when loading the youtils_test binary (observed with
Ubuntu 14.04's gdb 7.7.1 and a self compiled 7.12.1).
Independent of that, the outcome of these tests was that CapnProto RPC
is not a good fit for the intended use case, but the serialization is.

Part of https://github.com/openvstorage/volumedriver/issues/268