NetworkXioRequest: manage lifetime (+ attached future) with intrusive_ptr
Requests result in boost::futures returned from the async voldrv
interfaces. We don't really care about the result of the future as the
final continuation takes care of sending the response. For that the
returned future needs to be kept alive until this continuation is
done. This is done by managing the ownership/lifetime with a smart ptr
(specifically, an intrusive_ptr to allow interfacing with C / use of
intrusive containers).

Part of https://github.com/openvstorage/volumedriver/issues/268