volumedriver/failovercache: introduce new protocol + client + server
A new opcode is introduced to the old protocol. Messages with that
opcode follow a simpler pattern than the current protocol which is aimed
at streaming:
* a backward compatible header
* Capn'Proto message header
* optional bulk data
.

The existing server code is enhanced with a CapnProtoDispatcher that
handles messages with the new opcode.
The client code for the new protocol is newly introduced and exposes an
async interface built internally on boost::asio.

Part of https://github.com/openvstorage/volumedriver/issues/268