MDSMetaDataStore: check how far a slave is behind before promoting it
If there is a limit configured on the number of tlogs a slave can lag
behind (either in the volume's MDSMetaDataBackendConfig or globally via
'volume_manager/metadata_mds_slave_max_tlogs_behind'), a slave's
"uptodateness" is checked before promoting it to the master role. If
it's lagging too far behind it will not be promoted.

Part of https://github.com/openvstorage/volumedriver/issues/339