youtils: introduce MultiArchivePersistor
The idea is to migrate from serialization archives that are not forward
compatible (binary, text) to a forward compatible representation by
appending an XML archive right after the previous archive. New fields
are only used in the XML archive.
That way old client code can still (de)serialize the old
version (ignoring the trailing XML archive it is blissfully unaware of) and new
client code will try to deserialize the XML archive if present and use
that preferrably, otherwise falling back to the old version.

Part of https://github.com/openvstorage/volumedriver/issues/346